🔨 move component extraction code into it's own file
